Update nautilus-views.page for 3.8
Update instructions for the Views preferences. Reorder list items so
that they describe the preferences dialog first and main window second
because the introduction to the page is about the preferences dialog.

Use tooltip instead of image description for toolbar icons because Orca
reads out the tooltip instead of describing the button. Add note to
translators to use the tooltip translation.

Update defaults description for "Sort folders before files", bug 697051.

Delete quoted out sections of the page which described preferences which
are no longer in the UI.