Update help about automatic login
gnome-help: Create page about automatic login with instructions and link
to the sysadmin guide. Add to Makefile.am.

system-administrator-guide: Update the location of the config file to
match the upsteam default:
https://git.gnome.org/browse/gdm/tree/configure.ac#n187. Tidy up code
block and text. Delete autologin.page.stub as it has no content. Mark
strings that do not need translating. Move page from autologin.page to
user-autologin.page so that it matches the gnome-help page.