Review dconf custom defaults page, bug 694523
Link add seealso to dconf-lockdown because that is the next logical step
after setting defaults. Remove seealso to backgrounds-extra because that
is only one example of how to use the defaults.

Update revision status to review because the page is ready for review.

Improve wording, markup and indentation. Don't mention site.d because
this needs a page for itself or mention on a dconf landing page (TODO).
Remove external link as all the notes from that link should be covered
in the help and most already are.