Add more details on how to pair devices via Bluetooth
The page with more details is important because some of those features
are not known to end users ("cable pairing"), key combination are not
mentioned in the manual but in software ("BD remote") or multiple ways
to seemingly achieve the same thing are documented, but work in very
different ways ("Wiimote").

We might add more non-gaming devices to this list in the future.