Add classic-topbar-network-wired.svg and udpate topbar-night-light.svg
Looks like I missed an id attribute for classic-topbar-network-wired
in topbar-icons.svg. Added that back in. Fixes #134

Also, the night light icon was updated in 40.1, so I updated those
icons here to match.

Seems like the latest Inkscape exporter slightly changes the output,
so there's a bit of diff noise in the other exported images. Decided
to just go ahead and commit it so we don't carry the noise forever.